Switch env check to use ENV variable
Previously used the selected terraform state file environment to decide between prod or dev.

Two reasons to use our env var from deploy.sh:

1) Feels cleaner

Avoids setting some other state (the terraform state file environment)
for a different purpose and then working back from that.

2) Allows us to use the env var value in configuration

For instance, when setting bucket name we can use it for the bucket name
prefix. If we wanted to do this using the terraform env we'd be
restricted to having to make sure the terraform env always matches the
name of the env (coupling), whereas now we can continue to use a
'default' terraform env.

Feels acceptable to say "subdomain is always the name of the
environment". Feels dirtier to say "env is used to set terraform state
file configuration, then we use that to generate a subdomain".